# bench: Match ConnectBlock tx output counts