@monlovesmango 2025-04-03
Background for this PR can be found in https://github.com/bitcoin/bitcoin/pull/31689.

Happy to remove the renaming of `taproot_tx` if reviewers feel that it is inappropriate.