@monlovesmango 2025-04-03
> My output is less verbose compared to the outputs by other reviewers for some reason.

This is how it looks for me too when I run it on my mac, not sure why.