@davidgumberg 2025-04-04
Tested ACK 924f25f6fc766cb89a23fb

Nice find! Given that, as I understand, part of the motivation for this benchmark's introduction (#31689) is to be able to compare `ConnectBlock()` performance with and without schnorr batch validation (#29491), and part of the motivation for using Schnorr signatures in the first place is that they're susceptible to batch validation, it seems to me that it would be useful for these three benchmarks to be as like as possible, so that they can be compared, and this PR makes that the case in my testing.

Recent PR review club where this was discussed: (https://bitcoincore.reviews/31689#l-60). 

<details>

<summary>

## Benchmark Results 

</summary>

### master (c66f7dab33fb08dc323)

```console
$ git checkout --detach 924f25f~1
$ cmake -B build -DBUILD_BENCH=ON && cmake --build build -j $(nproc)
$ ./build/bin/bench_bitcoin --filter=ConnectBlock.* --min-time=5000
```

|            ns/block |             block/s |    err% |       ins/block |       cyc/block |    IPC |      bra/block |   miss% |     total | benchmark
|--------------------:|--------------------:|--------:|----------------:|----------------:|-------:|---------------:|--------:|----------:|:----------
|       46,549,571.00 |               21.48 |    1.2% |  672,393,930.64 |  161,287,432.73 |  4.169 |  15,122,140.18 |    1.7% |      5.46 | `ConnectBlockAllEcdsa`
|       47,199,469.60 |               21.19 |    0.2% |  680,249,499.27 |  164,132,414.09 |  4.145 |  15,120,173.64 |    1.7% |      5.47 | `ConnectBlockAllSchnorr`
|       59,705,754.22 |               16.75 |    0.7% |  839,091,363.50 |  207,657,368.33 |  4.041 |  18,945,901.12 |    1.8% |      5.55 | `ConnectBlockMixedEcdsaSchnorr`

### pr branch (924f25f6fc766cb)


```console
$ git checkout --detach 924f25f
$ cmake -B build -DBUILD_BENCH=ON && cmake --build build -j $(nproc)
$ ./build/bin/bench_bitcoin --filter=ConnectBlock.* --min-time=5000
```

|            ns/block |             block/s |    err% |       ins/block |       cyc/block |    IPC |      bra/block |   miss% |     total | benchmark
|--------------------:|--------------------:|--------:|----------------:|----------------:|-------:|---------------:|--------:|----------:|:----------
|       58,425,229.87 |               17.12 |    0.6% |  840,316,339.12 |  203,096,705.00 |  4.138 |  18,694,845.12 |    1.7% |      5.45 | `ConnectBlockAllEcdsa`
|       59,756,200.89 |               16.73 |    0.6% |  846,439,909.12 |  207,679,122.78 |  4.076 |  18,670,295.75 |    1.8% |      5.52 | `ConnectBlockAllSchnorr`
|       59,767,476.56 |               16.73 |    0.6% |  840,513,726.44 |  207,604,950.00 |  4.049 |  18,968,510.88 |    1.8% |      5.58 | `ConnectBlockMixedEcdsaSchnorr`

</details>

----

> My output is less verbose compared to the outputs by other reviewers for some reason.

This is because nanobench provides some extra stats via linux's `perf_events` subsystem and omits them when they're not available:

> CPU statistics like instructions, cycles, branches, branch misses are only available on Linux, through [perf events](http://web.eece.maine.edu/~vweaver/projects/perf_events/). On some systems you might need to [change permissions](https://www.kernel.org/doc/html/latest/admin-guide/perf-security.html#unprivileged-users) through perf_event_paranoid or use ACL.

https://github.com/martinus/nanobench/blob/e4327893194f06928012eb81cabc606c4e4791ac/src/docs/tutorial.rst?plain=1#L79-L84