@monlovesmango 2025-04-04
Hey @l0rinc ,

I appreciate you reiterating this feedback! I actually tend to agree with you, however do not quite agree with the strategy laid out in your suggested changes, and my lack of experience makes me hesitant to make any sweeping changes.

For your suggested changes, in order for the metric to be tx/s I believe the batch size would actually need to be the number of txs and not the outputs.size(). Additionally, I actually think the metric we would really want is tx outputs/s, as switching to tx/s wouldn't actually resolve the issue this pr focused on. From a technical perspective it would be inaccurate to say that this benchmark is actually measuring tx outputs/s just by scaling the batch because what it actually would bench is just connectblock divided by the number of tx outputs which doesn't change the metric in a meaningful way.

More broadly, for a "ConnectBlock" benchmark my assumption is that this could potentially be used to benchmark other aspects of connectblock functionality, not just signature validation, in which case tx outputs/s would not really be an appropriate metric. I do think it might be valuable to have a separate benchmark that is more focused at the tx output granularity level.

> I think it would clarify things if we defined more precisely what we're measuring.

In my view, block/s is the most precise definition of what we're measuring and tx/s (or tx output/x) would be an obfuscation.