@l0rinc 2025-04-04
> In my view, block/s is the most precise definition of what we're measuring and tx/s (or tx output/x) would be an obfuscation.

Isn't the point of this PR to change what blocks/s means? That's why this PR exists, since a block doesn't really narrow down what we're measuring. We might as well call it `thing/s`.

> the metric we would really want is tx outputs/s

I think that's basically what I wrote (you can scale it to actual transactions, but since that's the same for all, it's just a multiplier, won't change the ratios)

We can also write:
```C++
bench.batch(test_block.vtx.size() * outputs.size()).unit("txo").run([&] {
```
or even
```C++
bench.batch(std::accumulate(test_block.vtx.begin(), test_block.vtx.end(), size_t{0}, [](size_t sum, const auto& tx) { return sum + tx->vout.size(); })).unit("txo").run([&] {
```
if you think that's more accurate, but calling it "block/s" and than changing the value and still calling it "block/s" is not the sign of trustworthy benchmark.

Will let you decide, I basically already said these in the other PR as well.