@monlovesmango 2025-04-04
>Isn't the point of this PR to change what blocks/s means?

The point of this PR is simply to make the existing benchmark code more consistent across benchmarks, not to change anything in the design of the benchmark that has already been merged. As I said before, I do not have the experience nor expertise to tackle something like that and am not convinced that just changing the unit of the benchmark is meaningful.